requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
openpyxl==3.1.5
passlib[bcrypt]==1.7.4
aiosqlite==0.20.0
bcrypt==4.0.1
pandas>=2.0
openpyxl>=3.1

